LayoutTests/fast/events/popup-blocked-from-unique-frame-via-window-open-named-sibling-frame.html

<!DOCTYPE html><!-- webkit-test-runner [ JavaScriptCanOpenWindowsAutomatically=false ] -->
<html>
<head>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.setCanOpenWindows();
}
</script>
</head>
<body>
<iframe name="A"></iframe>
<iframe name="B" src="data:text/html,<script>console.log(window.open('about:blank', 'A') ? 'FAIL' : 'PASS');</script>"></iframe>
</body>
</html>
