LayoutTests/fast/events/popup-blocked-from-unique-frame-via-window-open-named-sibling-frame-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to initiate navigation for frame with URL 'about:blank' from frame with URL 'data:text/html,<script>alert(window.open('about:blank', 'A') ?%20%27FAIL%27%20:%20%27PASS%27);%3C/script%3E'. The frame attempting navigation is neither same-origin with the target, nor is it the target's parent or opener.

ALERT: PASS
 
